Moves util packages to separate folder for easier organization
