+++ src/main/java/com/exitium/whewheo/particles/util/ReflectionUtils.java
@@ -1,4 +1,4 @@
-package com.exitium.whewheo.particles;
+package com.exitium.whewheo.particles.util;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
